Menu in ner template
For performance menus been moved to ner

No need to create new subject as there is only one source of values
