+++ src/app/annotations/ner/token/token.component.scss
@@ -1,3 +1,9 @@
 .mat-menu .mat-menu-panel {
     max-height: calc(50vh - 50px) !important;
 }
+
+button {
+    margin-right: 8px;
+    margin-bottom: 8px;
+    height: 52px;
+}
